Script_Tool-cursor/requirements.txt
pyserial
pywin32
pywinauto
pillow

